pyproject.toml
[project]
name = "air-quality-prediction"
version = "0.1.0"
description = "Prediction of AQI based on weather and emissions"
readme = "README.md"
authors = [
    { name = "Ivan Makhrov", email = "ismakhrov@edu.hse.ru" }
]
requires-python = ">=3.10"
dependencies = [
    "dvc>=3.65.0",
    "hydra-core>=1.3.2",
    "lightning>=2.6.0",
    "mlflow>=3.8.0",
    "numpy==1.26.4",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "ruff>=0.14.9",
    "scikit-learn>=1.7.2",
    "tabpfn>=6.2.0",
    "torch<2.8",
]

[project.scripts]
air-quality-prediction = "air_quality_prediction.command:main"

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[tool.uv]
required-environments = [
  "sys_platform == 'darwin' and platform_machine == 'x86_64' and python_version >= '3.9'"
]

[dependency-groups]
dev = [
    "fire>=0.7.1",
    "pre-commit>=4.5.0",
]

[tool.ruff]
line-length = 100
format = { docstring-code-format = true }

[tool.ruff.lint]
select = [
    # isort
    "I",
    # pycodestyle
    "E", "W",
    # Pyflakes
    "F",
    # Pylint - very strict, may be use when have enough time
    "PL",
]

pycodestyle = { max-doc-length = 105 }

[tool.ruff.lint.pylint]
max-args = 10  # default is 5
